requirements.txt
torch
ultralytics
gradio
opencv-python
pyzbar
Pillow
